Move table is now generated if doesn't exist and loads properly,     still todo tests.

+++ solver/src/cubie.rs
@@ -173,22 +173,22 @@ impl CubieCube {
     }
 
     // Corner orientation coord, 0..2186, convert orientation in order to ternary number
-    pub fn get_twist(&self) -> u32 {
-        let mut total: u32 = 0;
+    pub fn get_twist(&self) -> u16 {
+        let mut total: u16 = 0;
         for i in 0..7 { // Ignore DBR since it can be calculated from others
-            total = 3 * total + self.co[i] as u32;
+            total = 3 * total + self.co[i] as u16;
         }
         total
     }
 
-    pub fn set_twist(&mut self, mut twist: u32) {
-        let mut tp: u32 = 0;
+    pub fn set_twist(&mut self, mut twist: u16) {
+        let mut tp: u16 = 0;
 
         for i in (0..7).rev() {
             // Have to -1 inside because usize can't be negative and the compiler doesn't
             // understand range is exclusive at the end
             self.co[i as usize] = (twist % 3) as i8;
-            tp += self.co[i as usize] as u32;
+            tp += self.co[i as usize] as u16;
             twist = twist / 3;
         }
